requirements.txt
autopep8==1.5.6
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.16.0
python-http-client==3.3.2
requests==2.25.1
sendgrid==6.6.0
six==1.15.0
SQLAlchemy==1.4.5
starkbank-ecdsa==1.1.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
